[Chore] Restructure the folder structure
