--- WeConnect_RestAPI/WeConnect_RestAPI/api/we_connect/custom_validation.py
@@ -1,11 +1,0 @@
-import re
-
-def is_empty(the_input):
-    if the_input.strip() == '':
-        return True
-    return False
-
-def is_email(the_input):
-    if not re.match(r"[^@]+@[^@]+\.[^@]+", the_input):
-        return True
-    return False
