+++ WeConnect_RestAPI/settings.py
